Updated ui for workspace Archival on the admin side
